BoolMixin is no longer needed
